+++ Editor/OptimizerSession.cs
@@ -10,8 +10,6 @@ namespace Anatawa12.AvatarOptimizer
     internal class OptimizerSession
     {
         private readonly GameObject _rootObject;
-        private readonly Dictionary<Object, Object> _mapping = new Dictionary<Object, Object>();
-        private readonly List<Object> _toDestroy = new List<Object>();
         private readonly HashSet<Object> _added = new HashSet<Object>();
         private readonly DummyObject _assetFileObject;
         public ObjectMappingBuilder MappingBuilder { get; }
@@ -28,11 +26,6 @@ namespace Anatawa12.AvatarOptimizer
             MappingBuilder = new ObjectMappingBuilder(rootObject);
         }
 
-        public void AddObjectMapping<T>(T oldValue, T newValue) where T : Object => _mapping[oldValue] = newValue;
-
-        public void Destroy(Object merge) => _toDestroy.Add(merge);
-        internal List<Object> GetObjectsToDestroy() => _toDestroy;
-
         public T GetRootComponent<T>() where T : Component
         {
             return _rootObject != null ? _rootObject.GetComponent<T>() : null;
@@ -41,8 +34,7 @@ namespace Anatawa12.AvatarOptimizer
         public IEnumerable<T> GetComponents<T>() where T : Component
         {
             return (_rootObject != null ? _rootObject.GetComponentsInChildren<T>(true) : Object.FindObjectsOfType<T>())
-                .Where(x => x)
-                .Where(x => !_toDestroy.Contains(x));
+                .Where(x => x);
         }
 
         public T AddToAsset<T>(T obj) where T : Object